feat(backend,web): 대시보드 기본현황 근무형태별을 환산인원 SUM으로 변경
기본 현황 탭 "근무형태별 고정/격고/순회 인원현황"을 조회 기준 월의
월별 여사원 통합일정(MFEIS) ConvertedHeadcount 합산으로 표시.

- WorkTypeStats 필드 Int(인원 카운트) → BigDecimal(환산인원 SUM)
- buildBasicStats: .size → sumHeadcount(...) (SF scale=4 정합 헬퍼 재사용)
- web 막대 차트 소수점 첫째 자리 표시(decimals 옵션) + 제목에 (환산인원) 명시
- 서비스 테스트에 환산인원 SUM 검증 케이스 추가, 컨트롤러 테스트 픽스처 타입 정합

+++ backend/src/main/kotlin/com/otoki/powersales/admin/dto/response/DashboardResponse.kt
@@ -87,8 +87,9 @@ data class AgeGroupCount(
     val count: Int
 )
 
+/** 근무형태(고정/격고/순회)별 환산인원 SUM — MFEIS ConvertedHeadcount__c(Number 18,4) 정합. */
 data class WorkTypeStats(
-    val fixed: Int,
-    val alternating: Int,
-    val visiting: Int
+    val fixed: BigDecimal,
+    val alternating: BigDecimal,
+    val visiting: BigDecimal
 )
